chore: move loader
