+++ src/routes/blog/index.svelte
@@ -1,10 +1,39 @@
 <script context="module" lang="ts">
-	export { load } from '$lib/routes/blog/list/loader';
+	import type { Load } from '@sveltejs/kit';
+
+	import type { BlogPostEntries } from '$lib/services/notion/blog/entries/types';
+
+	import type { BlogPostListProps } from '$lib/routes/blog/list/types';
+	import { buildOgImgUrl } from '$lib/utils/buildOgImgUrl';
+	import { BASE_URL } from '$lib/constants/baseUrl';
+
+	export const load: Load<Record<string, string>, BlogPostListProps> = async ({ fetch }) => {
+		const blogPosts = (await fetch('/posts').then(
+			async (res) => await res.json()
+		)) as BlogPostEntries;
+
+		return {
+			props: {
+				blogPosts
+			},
+			stuff: {
+				seo: {
+					title: 'Blog Post',
+					openGraph: {
+						images: [{ url: buildOgImgUrl({ heading: 'Blog Post', text: `${BASE_URL}/blog` }) }]
+					}
+				}
+			},
+			cache: {
+				maxage: 60
+			}
+		};
+	};
 </script>
 
 <script lang="ts">
 	import BlogPostList from '$lib/components/blog/list.svelte';
-	import type { BlogPostListProps } from '$lib/routes/blog/list/types';
+	// import type { BlogPostListProps } from '$lib/routes/blog/list/types';
 
 	export let blogPosts: BlogPostListProps['blogPosts'];
 </script>
